new: user: transform-splash parameter to choose among crop or cover
